Change Player to Farmer. Implement rollDice12 method.
